# torque-api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Torque API Reference",
    "description": "This page contains information about the Torque APIs and how to use it.<br><br>To be able to test the API methods, you will need an access token to be set in the Authorize section.If you got to this page from your Torque account, we already made it available for you, but you can switch to a different token as needed.<br>To get a different token, use the 'Get token' function available under the 'Access Tokens' section. After running the method, copy the access_token you get in the response and set it in the Authorize fieldas 'Bearer access_token'. For example: Bearer fqSWuw72BbUVFn8AbokF77GJ0r5KEn9MiZjLXF8kBwI.<br><br>The API can be accessed with your account endpoint as will be mentioned in the descriptions and examples or using https://qtorque.io as the prefix instead."
  },
  "paths": {
    "/api/accounts/{account}/login": {
      "post": {
        "tags": [
          "Access Tokens"
        ],
        "summary": "Get token",
        "description": "Creates an API session token. Token validity period is 7200 seconds.",
        "parameters": [
          {
            "name": "account",
            "in": "path",
            "description": "Account name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelForLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
    "/api/token/longtoken": {
      "post": {
        "tags": [
          "Access Tokens"
        ],
        "summary": "Get long token",
        "description": "Requests a long-lived access token. Token validity period is 2147483647 seconds.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/token/refresh": {
      "post": {
        "tags": [
          "Access Tokens"
        ],
        "summary": "Refresh token",
        "description": "Extends an existing session. Use the new access token returned in the method’s response to reauthorize the session.",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Refresh token is invalid or was not given",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [ ]
      }
    },
    "/api/token/revoke": {
      "post": {
        "tags": [
          "Access Tokens"
        ],
        "summary": "Revoke token",
        "description": "Cancels the current API session (nullifies the active access token)",
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Failed to revoke access or refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/token/revokelongtoken": {
      "post": {
        "tags": [
          "Access Tokens"
        ],
        "summary": "Revoke long token",
        "description": "Cancels the current long token API session (nullifies the active long token)",
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Failed to revoke long access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/cloudaccounts/computeservices/k8s/agent-deployment-yaml": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Get the K8s deployment YAML",
        "description": "Gets the Kubernetes agent deployment YAML file.",
        "parameters": [
          {
            "name": "cloudAccount",
            "in": "query",
            "description": "Cloud account containing the Kubernetes cluster",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "computeService",
            "in": "query",
            "description": "Kubernetes cluster name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sandboxNamespaces",
            "in": "query",
            "description": "Namespace defined in Kubernetes cluster",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Cloud account, compute service, or cluster not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Cloud not processed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/cloudaccounts/computeservices/k8s/renew-cluster-token": {
      "put": {
        "tags": [
          "Agent"
        ],
        "summary": "Renew cluster token",
        "description": "Renews expired cluster token.",
        "parameters": [
          {
            "name": "cloudAccount",
            "in": "query",
            "description": "Cloud account containing the Kubernetes cluster",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "computeService",
            "in": "query",
            "description": "Kubernetes cluster name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Cloud account, compute service, or cluster not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Could not processed request"
          }
        }
      }
    },
    "/api/settings/computeservices/eks/agent-deployment-yaml": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Get EKS deployment YAML",
        "description": "Gets the Amazon Elastic Kubernetes service (eks) agent deployment YAML file.",
        "parameters": [
          {
            "name": "cloudAccount",
            "in": "query",
            "description": "Cloud account containing the Kubernetes cluster",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "computeService",
            "in": "query",
            "description": "Kubernetes cluster name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sandboxNamespace",
            "in": "query",
            "description": "Namespace defined in the Kubernetes cluster",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "roleArn",
            "in": "query",
            "description": "Rule ARN of the EKS",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Cloud account, compute service, or cluster not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Cloud not processed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Error name",
            "nullable": true
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Error information",
        "x-internal": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            },
            "description": "Array of errors",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Errors received from an API",
        "x-internal": true
      },
...
      "ModelForLogin": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "description": "User password"
          }
        },
        "additionalProperties": false,
        "description": "Login request",
        "x-internal": true
      },
...
      "RefreshTokenModel": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh token value",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Model of refresh_token returned in the response of login method",
        "x-internal": true
      },
...
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Access token string as issued by the authorization server",
            "nullable": true
          },
          "refresh_token": {
            "type": "string",
            "description": "Refresh token to be used to extend existing session",
            "nullable": true
          },
          "token_type": {
            "type": "string",
            "description": "Token type, typically just the string \"bearer\"",
            "nullable": true
          },
          "expires_in": {
            "type": "integer",
            "description": "Access token validity period",
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "description": "Describes the authentication server response if the request for an access token is valid\r\n<remarks>\r\n Contains access token and some additional properties about the authorization\r\n</remarks>",
        "x-internal": true
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}